Room back-end refactoring
